Revamped Javawolf. More refactoring.
All members of Javawolf have been changed to non-static. Javawolf configuration file is now loaded through the constructor. Connecting to server is now done in the main() method. Added parsing for server notices in Javawolf so that they are not logged as errors. Static references to Javawolf in WolfGame have been removed and replaced with instance calls because the instance of Javawolf is now passed into the instance of Wolfgame. Added accessor methods in WolfPlayer to simplify retrieving a player's role. Added village and wolfTeam ArrayLists in WolfGame to improve tracking of players.