Added a "votes" command.
Added votes command to display number of votes during the day. Added notice() method and changed privmsg() to msg() in WolfGame. Game logic seems to be broken at the moment. isNight property is not changing to false at the end of a night. Player counting also seems to be off. No compile errors.