Fixed bug in getPlayer(int).
Added check for negative index parameter in getPlayer(int) which was throwing index out of range exceptions. This was the source of the buggy behaviour seen in the last commit. Removed isAlive property of WolfPlayer and changed it to a dynamic check in WolfGame. It now checks if a player is either in the villagers or wolfTeam list. Since players are removed from those lists upon death, a player's presense in either list shows that they are alive. Changed some "msg" commands into "notice" commands.