More refactoring.
Added set methods in WolfPlayer for all the roles. Rearranged methods in WolfGame so all player action methods are grouped together. Added a delay in fquit to ensure players are devoiced. Added JavaWolf as the "real name" of the bot. Made a change to force player action methods to always log when they are called.