Added buggy "fjoin" command. More refactoring.
The bug with "fjoin" is the inability to dynamically get a player's IRC username and host. Sending a raw WHO command is insufficient. Moved all initialization of member variables in Javawolf into the constructor. Changed "login_str" to "password" and moved NickServ identification to onConnect() and now takes advantage of the identify() method. Changed logging to use logEvent() of the Javawolf instance where possible. Other small changes.